requirements.txt
streamlit==1.44.1
pandas==2.2.3
numpy==2.2.4
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
python-dotenv==1.1.0
APScheduler==3.10.4
